perf: properly cache createClientConfig in both prod & dev. It now only runs once, and does not re-run when refreshing the page or navigating. HMR still works despite caching in dev
