+++ packages/ui/src/utilities/buildTableState.ts
@@ -1,48 +1,20 @@
-import type { I18nClient } from '@payloadcms/translations'
 import type {
   BuildTableStateArgs,
   ClientCollectionConfig,
   ClientConfig,
   ErrorResult,
-  ImportMap,
   PaginatedDocs,
   SanitizedCollectionConfig,
-  SanitizedConfig,
 } from 'payload'
 
 import { dequal } from 'dequal' // TODO: Can we change this to dequal/lite ? If not, please add comment explaining why
-import { createClientConfig, formatErrors } from 'payload'
+import { formatErrors } from 'payload'
 
 import type { Column } from '../elements/Table/index.js'
 import type { ListPreferences } from '../elements/TableColumns/index.js'
 
+import { getClientConfig } from './getClientConfig.js'
 import { renderFilters, renderTable } from './renderTable.js'
-
-let cachedClientConfig = global._payload_clientConfig
-
-if (!cachedClientConfig) {
-  cachedClientConfig = global._payload_clientConfig = null
-}
-
-export const getClientConfig = (args: {
-  config: SanitizedConfig
-  i18n: I18nClient
-  importMap: ImportMap
-}): ClientConfig => {
-  const { config, i18n, importMap } = args
-
-  if (cachedClientConfig && process.env.NODE_ENV !== 'development') {
-    return cachedClientConfig
-  }
-
-  cachedClientConfig = createClientConfig({
-    config,
-    i18n,
-    importMap,
-  })
-
-  return cachedClientConfig
-}
 
 type BuildTableStateSuccessResult = {
   clientConfig?: ClientConfig
